requirements.txt
telethon==1.36.0
openai>=1.0.0
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
